fix: fs usage in browser
