+++ typescript/aleo-sdk/src/runtime.ts
@@ -4,8 +4,6 @@ export {
   toKeyId,
 } from './utils/helper.js';
 
-export { AleoProtocolProvider } from './clients/protocol.js';
 export { AleoProvider } from './clients/provider.js';
-export { AleoSigner } from './clients/signer.js';
 
 export type { AleoReceipt, AleoTransaction } from './utils/types.js';
